+++ src/oop/special_class_method/hard/q9/Earth.java
@@ -1,4 +1,4 @@
-package oop.variable_scope.medium.q1;
+package oop.special_class_method.hard.q9;
 
 public class Earth {
     private double landMass;
